refactor: move unknown related to its own folder
