+++ src/index.ts
@@ -3,9 +3,7 @@ export { required, requiredDeep } from 'unpartial'
 export * from './any-types'
 export * from './array'
 export * from './assertType'
-export * from './assertUnknown'
 export * from './assignability'
-export * from './checkUnknown'
 export * from './conditional'
 export * from './facade'
 export * from './function'
@@ -34,6 +32,7 @@ export * from './typeOverrideIncompatible'
 export * from './types/optional'
 export * from './types/required'
 export * from './UnionKeys'
+export * from './unknown'
 export * from './ValueOf'
 export { types, types as T }
