+++ src/unknown/index.ts
@@ -1,0 +1,2 @@
+export * from './assertUnknown'
+export * from './checkUnknown'
